+++ statsmodels/sandbox/gam_gsoc2015/gam.py
@@ -6,12 +6,11 @@ __date__ = '08/07/15'
 
 import numpy as np
 import scipy as sp
-from scipy.linalg import block_diag
 from statsmodels.discrete.discrete_model import Logit
 from scipy.stats import chi2
 from statsmodels.genmod.generalized_linear_model import GLM, GLMResults, GLMResultsWrapper, lm, _check_convergence
 from statsmodels.tools.sm_exceptions import PerfectSeparationError
-
+from statsmodels.sandbox.gam_gsoc2015.gam_penalties import UnivariateGamPenalty, MultivariateGamPenalty
 
 ## this class will be later removed and taken from another push
 class PenalizedMixin(object):
@@ -22,13 +21,14 @@ class PenalizedMixin(object):
 
     def __init__(self, *args, **kwds):
         super(PenalizedMixin, self).__init__(*args, **kwds)
-
+        print(kwds)
         penal = kwds.pop('penal', None)
         # I keep the following instead of adding default in pop for future changes
         if penal is None:
-            # TODO: switch to unpenalized by default
-            self.penal = None
+           print("Define a penalty")
+
         else:
+            print('the penal is ', penal, 'with alpha=', penal.alpha)
             self.penal = penal
 
         # TODO: define pen_weight as average pen_weight? i.e. per observation
@@ -146,196 +146,6 @@ class PenalizedMixin(object):
                 return res
 
 
-## this class will be later removed and taken from another push
-class Penalty(object):
-    """
-    A class for representing a scalar-value penalty.
-    Parameters
-    wts : array-like
-        A vector of weights that determines the weight of the penalty
-        for each parameter.
-    Notes
-    -----
-    The class has a member called `alpha` that scales the weights.
-    """
-
-    def __init__(self, wts):
-        self.wts = wts
-        self.alpha = 1.
-
-    def func(self, params):
-        """
-        A penalty function on a vector of parameters.
-        Parameters
-        ----------
-        params : array-like
-            A vector of parameters.
-        Returns
-        -------
-        A scalar penaty value; greater values imply greater
-        penalization.
-        """
-        raise NotImplementedError
-
-    def grad(self, params):
-        """
-        The gradient of a penalty function.
-        Parameters
-        ----------
-        params : array-like
-            A vector of parameters
-        Returns
-        -------
-        The gradient of the penalty with respect to each element in
-        `params`.
-        """
-        raise NotImplementedError
-
-
-class UnivariateGamPenalty(Penalty):
-    __doc__ = """
-    Penalty for Generalized Additive Models class
-
-    Parameters
-    -----------
-    alpha : float
-        the penalty term
-
-    wts: TODO: I do not know!
-
-    cov_der2: the covariance matrix of the second derivative of the basis matrix
-
-    der2: The second derivative of the basis function
-
-    Attributes
-    -----------
-    alpha : float
-        the penalty term
-
-    wts: TODO: I do not know!
-
-    cov_der2: the covariance matrix of the second derivative of the basis matrix
-
-    der2: The second derivative of the basis function
-
-    n_samples: The number of samples used during the estimation
-
-
-
-    """
-
-    def __init__(self, univariate_smoother, wts=1, alpha=1):
-
-        self.wts = wts #should we keep wts????
-        self.alpha = alpha
-        self.univariate_smoother = univariate_smoother
-        self.n_samples, self.n_columns = self.univariate_smoother.n_samples, self.univariate_smoother.dim_basis
-
-    def func(self, params):
-        '''
-        1) params are the coefficients in the regression model
-        2) der2  is the second derivative of the splines basis
-        '''
-
-        # The second derivative of the estimated regression function
-        f = np.dot(self.univariate_smoother.der2_basis_, params)
-
-        return self.alpha * np.sum(f**2) / self.n_samples
-
-    def grad(self, params):
-        '''
-        1) params are the coefficients in the regression model
-        2) der2  is the second derivative of the splines basis
-        3) cov_der2 is obtained as np.dot(der2.T, der2)
-        '''
-
-        return 2 * self.alpha * np.dot(self.univariate_smoother.cov_der2_, params) / self.n_samples
-
-    def deriv2(self, params):
-        return 2 * self.alpha * self.univariate_smoother.cov_der2_ / self.n_samples
-
-
-class MultivariateGamPenalty(Penalty):
-    __doc__ = """
-    GAM penalty for multivariate regression
-
-    Parameters
-    -----------
-    cov_der2: list of matrices
-     is a list of squared matrix of shape (size_base, size_base)
-
-    der2: list of matrices
-     is a list of matrix of shape (n_samples, size_base)
-
-    alpha: array-like
-     list of doubles. Each one representing the penalty
-          for each function
-
-    wts: array-like
-     is a list of doubles of the same length of alpha
-
-    """
-
-    def __init__(self, multivariate_smoother, alphas, wts=None):
-
-        if len(multivariate_smoother.smoothers_) != len(alphas):
-            raise ValueError('all the input values should be list of the same length. len(smoothers_)=', len(multivariate_smoother.smoothers_), ' len(alphas)=', len(alphas))
-
-        self.multivariate_smoother = multivariate_smoother
-        self.k_columns = self.multivariate_smoother.k_columns
-        self.k_variables = self.multivariate_smoother.k_variables
-        self.n_samples = self.multivariate_smoother.n_samples
-        self.alphas = alphas
-
-        # TODO: Review this
-        if wts is None:
-            self.wts = [1] * len(alphas)
-        else:
-            self.wts = wts
-
-        self.mask = [np.array([False]*self.k_columns)
-                     for _ in range(self.k_variables)]
-        param_count = 0
-        for i, smoother in enumerate(self.multivariate_smoother.smoothers_):
-
-            # the mask[i] contains a vector of length k_columns. The index
-            # corresponding to the i-th input variable are set to True.
-            self.mask[i][param_count: param_count + smoother.dim_basis] = True
-            param_count += smoother.dim_basis
-
-        self.gp = []
-        for i in range(self.k_variables):
-            gp = UnivariateGamPenalty(wts=self.wts[i], alpha=self.alphas[i],
-                                      univariate_smoother=self.multivariate_smoother.smoothers_[i])
-            self.gp.append(gp)
-
-        return
-
-    def func(self, params):
-        cost = 0
-        for i in range(self.k_variables):
-            params_i = params[self.mask[i]]
-            cost += self.gp[i].func(params_i)
-
-        return cost
-
-    def grad(self, params):
-        grad = []
-        for i in range(self.k_variables):
-            params_i = params[self.mask[i]]
-            grad.append(self.gp[i].grad(params_i))
-
-        return np.concatenate(grad)
-
-    def deriv2(self, params):
-        deriv2 = np.empty(shape=(0,0))
-        for i in range(self.k_variables):
-            params_i = params[self.mask[i]]
-            deriv2 = block_diag(deriv2, self.gp[i].deriv2(params_i))
-
-        return deriv2
-
-
 class LogitGam(PenalizedMixin, Logit):
     pass
 
@@ -418,15 +228,17 @@ class GLMGam(PenalizedMixin, GLM):
                  missing='none', **kwargs):
 
         gp = UnivariateGamPenalty(alpha=alpha, univariate_smoother=smoother)
-        super(GLMGam, self).__init__(endog, smoother.basis_, family=None, offset=None, exposure=None,
-                                     missing='none', penal=gp, **kwargs)
+        kwargs['penal'] = gp
+
+        super(GLMGam, self).__init__(endog, smoother.basis_, family=family, offset=offset, exposure=exposure,
+                                     missing=missing, **kwargs)
         self.smoother = smoother
 
         return
 
     def fit(self, start_params=None, maxiter=100, method='IRLS', tol=1e-8,
             scale=None, cov_type='nonrobust', cov_kwds=None, use_t=None,
-            full_output=True, disp=False, max_start_irls=3, **kwargs):
+            full_output=True, disp=False, max_start_irls=3, weights=None, **kwargs):
 
         if method.lower() == 'pirls':
 
@@ -437,13 +249,93 @@ class GLMGam(PenalizedMixin, GLM):
 
             return self._fit_pirls(self.endog, self.smoother.basis_, pen_matrix, self.alpha,
                                    start_params=None, maxiter=100, tol=1e-8,
-                                   scale=None, cov_type='nonrobust', cov_kwds=None, use_t=None, weights=None)
+                                   scale=scale, cov_type=cov_type, cov_kwds=cov_kwds, use_t=use_t, weights=weights)
 
         else:
             return super(GLMGam, self).fit(start_params=start_params, maxiter=maxiter, method=method, tol=tol,
-                                                   scale=scale, cov_type=cov_type, cov_kwds=cov_kwds, use_t=use_t,
-                                                   full_output=full_output, disp=disp, max_start_irls=max_start_irls,
-                                                   **kwargs)
+                                           scale=scale, cov_type=cov_type, cov_kwds=cov_kwds, use_t=use_t,
+                                           full_output=full_output, disp=disp, max_start_irls=max_start_irls,
+                                           **kwargs)
+
+    # pag 165 4.3 # pag 136 PIRLS
+    def _fit_pirls(self, y, spl_x, spl_s, alpha, start_params=None, maxiter=100, tol=1e-8,
+                   scale=None, cov_type='nonrobust', cov_kwds=None, use_t=None, weights=None):
+
+        endog = y
+        wlsexog = spl_x
+
+        n_samples, n_columns = wlsexog.shape
+
+        # TODO what are these values?
+        if weights is None:
+            self.data_weights = np.array([1.] * n_samples)
+        else:
+            self.data_weights = weights
+
+        self._offset_exposure = np.array([.1] * n_samples)
+        self.scaletype = 'dev'
+
+        if start_params is None:
+            mu = self.family.starting_mu(endog)
+            lin_pred = self.family.predict(mu)
+        else:
+            lin_pred = np.dot(wlsexog, start_params) + self._offset_exposure
+            mu = self.family.fitted(lin_pred)
+        dev = self.family.deviance(endog, mu)
+
+        history = dict(params=[None, start_params], deviance=[np.inf, dev])
+        converged = False
+        criterion = history['deviance']
+        # This special case is used to get the likelihood for a specific
+        # params vector.
+        if maxiter == 0:
+            mu = self.family.fitted(lin_pred)
+            self.scale = self.estimate_scale(mu)
+            wls_results = lm.RegressionResults(self, start_params, None)
+            iteration = 0
+
+        for iteration in range(maxiter):
+
+            # TODO: is this equivalent to point 1 of page 136: w = 1 / (V(mu) * g'(mu))  ?
+            self.weights = self.data_weights * self.family.weights(mu)
+
+            #TODO: is this equivalent to point 1 of page 136:  z = g(mu)(y - mu) + X beta  ?
+            wlsendog = (lin_pred + self.family.link.deriv(mu) * (endog-mu)
+                        - self._offset_exposure)
+
+            # this defines the augmented matrix point 2a on page 136
+            wls_results = penalized_wls(wlsexog, wlsendog, spl_s, self.weights, alpha)
+            lin_pred = np.dot(wlsexog, wls_results.params).ravel() + self._offset_exposure
+            mu = self.family.fitted(lin_pred)
+
+            history = self._update_history(wls_results, mu, history)
+
+            self.scale = self.estimate_scale(mu)
+            if endog.squeeze().ndim == 1 and np.allclose(mu - endog, 0):
+                msg = "Perfect separation detected, results not available"
+                raise PerfectSeparationError(msg)
+            converged = _check_convergence(criterion, iteration, tol)
+            if converged:
+                break
+        self.mu = mu
+
+        glm_results = GLMResults(self, wls_results.params,
+                                 wls_results.normalized_cov_params,
+                                 self.scale,
+                                 cov_type=cov_type, cov_kwds=cov_kwds,
+                                 use_t=use_t)
+
+        glm_results.method = "PIRLS"
+        history['iteration'] = iteration + 1
+        glm_results.fit_history = history
+        glm_results.converged = converged
+        return GLMResultsWrapper(glm_results)
+
+
+# TODO: This is an old version of GLMGAM and is used only for testing. This will soon be removed
+class OLD_GLMGam(PenalizedMixin, GLM):
+
+    _results_class = GLMGAMResults
 
     # pag 165 4.3 # pag 136 PIRLS
     def _fit_pirls(self, y, spl_x, spl_s, alpha, start_params=None, maxiter=100, tol=1e-8,
